backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.0
bcrypt==4.0.1
argon2-cffi==23.1.0
python-decouple==3.8
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Redis
redis==5.0.1

# Celery
celery==5.3.6

# MinIO
minio==7.2.5

# Weaviate
weaviate-client==4.16.7

# Neo4j
neo4j==5.15.0

# Additional dependencies
httpx==0.26.0
aiofiles==23.2.1
Pillow==10.2.0